Moved datatypes to separate file and some terminal functions to another
